Explicit -> implicit iteration.
This eliminates the whole class of bugs where the iteration order is
inconsistent with the data model.  It also simplifies the code by simply
removing the whole 9-nested-layers of for loops along with the barnacles
that were growing there (multiple passes for selling and then buying,
different iteration order depending on whether the planet sells Eden
warps, etc.)

The downside: Parallelism is lost.  This is now a single-threaded
program.